+++ docs/source/releases.rst
@@ -19,6 +19,7 @@ Additions
 - Added :attr:`Message.referenced_message` attribute.
 - Added module ``qord.utils``, see :ref:`reference-utilities` for more info.
 - Added :func:`qord.utils.create_timestamp` helper function.
+- Added equality comparison support for various Discord models.
 
 Fixes
 ~~~~~